requirements.txt
alembic==1.16.5
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.10.0
bcrypt==4.3.0
bidict==0.23.1
black==25.9.0
blinker==1.9.0
Brotli==1.1.0
cached-property==2.0.1
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
ConfigArgParse==1.7.1
cryptography==46.0.1
db==0.1.1
db-sqlite3==0.0.1
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.14
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.41.1
locust-cloud==1.27.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.8.0
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-gettext==5.0
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pytokens==0.1.10
PyYAML==6.0.2
pyzmq==27.1.0
redis==6.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
ruff==0.13.1
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
standard==1.0.3
starlette==0.47.3
termcolor==3.1.0
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.0
zope.interface==8.0.1
